Implementert en del kunst

+++ Assets/Scripts/Player/Weapon System/WeaponStats.cs
@@ -6,6 +6,6 @@ public class WeaponStats : ScriptableObject
     public float fireRate;
     public float bulletSpeed;
     public float breathUsage;
-    public int weaponDamage;
-    public int swapTime; 
+    public float weaponDamage;
+    public float swapTime; 
 }
